Rework app-image updater
