+++ src/auto-updater/bfx.appimage.updater.js
@@ -1,57 +1,15 @@
 'use strict'
 
-const path = require('path')
-const fs = require('fs')
 const {
   AppImageUpdater
 } = require('electron-updater')
-const log = require('electron-log')
 
-const appDir = path.dirname(require.main.filename)
 const {
-  version
-} = require(path.join(appDir, 'package.json'))
+  prepareInstall,
+  setAppImagePathIfZipRelease
+} = require('./utils')
 
-const isZipRelease = (root) => {
-  const isZipReleaseFile = path.join(root, 'isZipRelease')
-
-  return fs.existsSync(isZipReleaseFile)
-}
-
-const root = path.join(appDir, '../..')
-
-if (isZipRelease(root)) {
-  process.env.APPIMAGE = path.join(
-    root,
-    '..',
-    `BitfinexReport-${version}-x64-linux.AppImage`
-  )
-}
-
-const makeTempReleaseFile = (root) => {
-  const fileName = `BitfinexReport-${version}-x64-linux.AppImage`
-  const filePath = path.join(root, '..', fileName)
-
-  fs.writeFileSync(filePath, '')
-
-  return filePath
-}
-
-const prepareInstall = () => {
-  const root = path.join(appDir, '../..')
-
-  if (!isZipRelease(root)) {
-    return
-  }
-
-  try {
-    fs.rmdirSync(root, { recursive: true })
-  } catch (err) {
-    log.error(err)
-  }
-
-  process.env.APPIMAGE = makeTempReleaseFile(root)
-}
+setAppImagePathIfZipRelease()
 
 class BfxAppImageUpdater extends AppImageUpdater {
   doInstall (...args) {
