+++ src/auto-updater/utils.js
@@ -1,0 +1,68 @@
+'use strict'
+
+const path = require('path')
+const fs = require('fs')
+const log = require('electron-log')
+
+const appDir = path.dirname(require.main.filename)
+const _root = path.join(appDir, '../..')
+const {
+  version
+} = require(path.join(appDir, 'package.json'))
+
+const isZipRelease = (root = _root) => {
+  const isZipReleaseFile = path.join(root, 'isZipRelease')
+
+  return fs.existsSync(isZipReleaseFile)
+}
+
+const getTempReleaseFilePath = (root = _root) => {
+  const fileName = `BitfinexReport-${version}-x64-linux.AppImage`
+  const filePath = path.join(root, '..', fileName)
+
+  return filePath
+}
+
+const setAppImagePath = (
+  appImagePath = getTempReleaseFilePath()
+) => {
+  process.env.APPIMAGE = appImagePath
+}
+
+const setAppImagePathIfZipRelease = () => {
+  if (!isZipRelease()) return
+
+  setAppImagePath()
+}
+
+if (isZipRelease()) {
+  setAppImagePath()
+}
+
+const makeTempReleaseFile = (filePath) => {
+  fs.writeFileSync(filePath, '')
+}
+
+const prepareInstall = () => {
+  const root = path.join(appDir, '../..')
+
+  if (!isZipRelease(root)) {
+    return
+  }
+
+  try {
+    fs.rmdirSync(root, { recursive: true })
+  } catch (err) {
+    log.error(err)
+  }
+
+  const filePath = getTempReleaseFilePath(root)
+  makeTempReleaseFile(filePath)
+  setAppImagePath(filePath)
+}
+
+module.exports = {
+  isZipRelease,
+  prepareInstall,
+  setAppImagePathIfZipRelease
+}
